server/requirements.txt
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.9.0
attrs==25.3.0
awsebcli==3.24.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blessed==1.21.0
blinker==1.9.0
botocore==1.38.24
cachetools==5.5.2
cement==2.10.14
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
clr_loader==0.2.7.post0
colorama==0.4.6
cryptography==45.0.3
dataclasses-json==0.6.7
decorator==5.2.1
deepgram-sdk==4.1.0
Deprecated==1.2.18
deprecation==2.1.0
docopt==0.6.2
fabric==3.2.2
Flask==3.1.1
flask-cors==6.0.0
frozenlist==1.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-genai==1.16.1
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
gTTS==2.5.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
invoke==2.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
jinxed==1.3.0
jmespath==1.0.1
Js2Py==0.74
lxml==5.4.0
MarkupSafe==3.0.2
marshmallow==3.26.1
MouseInfo==0.1.3
multidict==6.4.4
mypy_extensions==1.1.0
packaging==24.2
paramiko==3.5.1
pathspec==0.12.1
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
pipwin==0.5.2
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio @ https://abstractumbra.github.io/pip/PyAudio/PyAudio-0.2.14-cp312-cp312-win_amd64.whl#sha256=12f2f1ba04e06ff95d80700a78967897a489c05e093e3bffa05a84ed9c0a7fa3
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pygame==2.6.1
PyGetWindow==0.0.9
pyjsparser==2.7.1
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyparsing==3.2.3
pypdfium2==4.30.1
pyperclip==1.9.0
pypiwin32==223
PyPrind==2.11.3
PyRect==0.2.0
PyScreeze==1.0.1
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
pythonnet==3.0.5
pytweening==1.2.0
pywin32==310
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
semantic-version==2.10.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
termcolor==2.5.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==1.26.20
wcwidth==0.2.13
webrtcvad==2.0.10
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.0
